DC#38: Max Call Stack Size Exceeded
This issue is occurring because one player's king is able to castle and
the other player has selected a piece that has a valid move such that
their king would be able to castle, and the evaluation of valid moves is
causing one king to determine if they can castle which subsequently
causes the other king to determine if they can castle, and so on until
the maximum call stack size is exceeded.

Fixed by checking whether castling is possible only when the king is
selected rather than all times the king is asked for valid moves.